app/requirements.txt
flask
gunicorn
mysql-connector-python
# python-dotenv